requirements.txt
asgiref==3.3.4
Django==3.2.3
django-3-jet==1.0.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
